add IEmbeddingVar interface
